+++ test/Boost.t.sol
@@ -11,8 +11,8 @@ abstract contract BoostTest is Test, GasSnapshot {
     event Claim(IBoost.ClaimConfig claim);
     event Deposit(uint256 boostId, address sender, uint256 amount);
     event Burn(uint256 boostId);
-    event EthFeeSet(uint128 ethFee);
-    event TokenFeeSet(uint16 tokenFee);
+    event EthFeeSet(uint256 ethFee);
+    event TokenFeeSet(uint256 tokenFee);
     event EthFeesCollected(address recipient);
     event TokenFeesCollected(IERC20 token, address recipient);
 
@@ -80,7 +80,7 @@ abstract contract BoostTest is Test, GasSnapshot {
         address _guard,
         uint256 _start,
         uint256 _end,
-        uint128 _ethFee
+        uint256 _ethFee
     ) internal returns (uint256) {
         uint256 boostID = boost.nextBoostId();
         vm.prank(_owner);
